requirements.txt
# NOTE: Please try and keep this file tidy.
# Only add direct requirements here. Any sub-dependencies will be set through
# the `freeze-dependencies` make target.
# To update dependencies, optionally update this file and then
# run the `update-dependencies` make target, and

# Django packages
django
django-bootstrap-form
django-crispy-forms
django-dotenv
django-prosemirror
django-registration-redux
django-termsandconditions
django-utils-six

# Misc
pillow
commonmark
diff_match_patch
lxml
markdown2
pilkit
python-dateutil
python-graph-core
requests
signxml
django_mdmail
setuptools
